Use the catalog parser
