+++ project/Settings.scala
@@ -18,6 +18,7 @@ object Settings {
     val disciplineMUnit  = "1.0.9"
     val fs2              = "3.3.0"
     val fs2Data          = "1.6.0"
+    val fs2Dom           = "0.0-9abbf5e-SNAPSHOT"
     val geminiLocales    = "0.7.0"
     val http4s           = "0.23.16"
     val http4sDom        = "0.2.3"
@@ -144,6 +145,12 @@ object Settings {
       deps(
         "co.fs2" %%% "fs2-node"
       )(fs2)
+    )
+
+    val FS2Dom = Def.setting(
+      Seq(
+        "com.armanbilge" %%% "fs2-dom" % fs2Dom
+      )
     )
 
     val GeminiLocales = Def.setting(
